# Cutting the SignalProducer overhead further.