--- a/Sources/Disposable.swift
+++ b/Sources/Disposable.swift
@@ -38,9 +38,27 @@ extension UnsafeAtomicState where State == DisposableState {
 	}
 }
 
+/// A disposable that does not have side effect upon disposal.
+internal final class _SimpleDisposable: Disposable {
+	private let state = UnsafeAtomicState<DisposableState>(.active)
+
+	var isDisposed: Bool {
+		return state.is(.disposed)
+	}
+
+	func dispose() {
+		_ = state.tryDispose()
+	}
+
+	deinit {
+		state.deinitialize()
+	}
+}
+
+/// A disposable that has already been disposed.
 internal final class NopDisposable: Disposable {
 	static let shared = NopDisposable()
-	var isDisposed = false
+	var isDisposed = true
 	func dispose() {}
 	private init() {}
 }