### Goals
Taking advantage of the deferred nature of `SignalProducer` to further reduce the overhead.

### TL;DR
1. Composing `SignalProducer` is going to be _way, way_ cheaper than composing `Signal`.

1. Constant `SignalProducer`s, e.g. `init(value:)` and sequence `init(_:)`, are going to do _significantly_ less work.

1. There is still a ~16x overhead over standard library `lazy` for purely collection manipulation. But given the type erasing and reactive (passive) nature of FRP/ReactiveSwift, there is probably very little stuff we can do about it. IOW optimizing for this is a non-goal. :)

![yeah](https://thumbs.gfycat.com/NecessaryIllinformedEnglishpointer-size_restricted.gif)

The actual speedup is subject to the new internal API adoption — unmigrated `SignalProducer` operators would still be relying on the conventional behavior and/or `lift`.

For example, if you do:

```swift
SignalProducer(Array(repeating: 1, count: 1000)).skipRepeats()
```

The performance remains the same as `master`, because `skipRepeats` hasn't been migrated yet.

### Introducing `SignalProducerCore`
`SignalProducerCore` is the new internal abstraction of `SignalProducer`. It has subsumed three major responsibilities from `SignalProducer`:

1. ~~the actual implementation of `startWithSignal`;~~
1. the actual implementation of `start` which accepts an `Signal.Observer`; and
1. the builder closure, now as `Core.make()`, which produces a `Signal` and the relevant context for composition and `startWithSignal`.

`SignalProducerCore` provides default implementations of all these capabilities for its subclasses, except for `Core.make()`.

_The cores are not nested in `SignalProducer` because the Swift 3.1 compiler would get stuck. The Swift 4 compiler has resolved the issue though._

### `SignalProducerCore` subclasses
#### `SignalCore`
This is the conventional `SignalProducer` with the `Signal` operator lifting optimisation in ReactiveSwift 2.0 (#140).

#### `EventTransformingCore`
`EventTransformingCore` composes _event transforms_, and is exposed as a new operator `Core.flatMapEvent` (internal for now).

It takes advantage of the deferred, single-observer nature of `SignalProducer`. For example, when we do:

```swift
upstream.map(transform).filterMap(filteringTransform).start()
```

It is contractually guaranteed that these operators would always end up producing a chain of streams, each with a **single and persistent** observer to its upstream. The multicasting & detaching capabilities of `Signal` is useless in these scenarios.

So `EventTransformingCore` builds on top of this very fact, and composes directly at the level of event transforms. This change implicitly nullifies #129, since producers no longer use `Signal` unless it absolutely has to.

It is intended for all synchronous `SignalProducer` operators. To encourage code reusing, the pilot operators have their implementation moved under `Signal.Event`, and `Signal` has gained `flatMapEvent` to use these `Event`-level operator implementations.

This represents a new type of `SignalProducer` that **does not use `Signal`** unless it is requested via `make()`.

##### Example
```swift
extension Signal {
	public func filterMap<U>(_ transform: @escaping (Value) -> U?) -> Signal<U, Error> {
		return flatMapEvent(Signal.Event. filterMap(transform))
	}
}

extension SignalProducer {
	public func filterMap<U>(_ transform: @escaping (Value) -> U?) -> SignalProducer<U, Error> {
		return core.flatMapEvent(Signal.Event. filterMap(transform))
	}
}
``` 

#### `EventGeneratingCore`
`EventGeneratingCore` wraps a generator closure that would be invoked when started. All the generator closure have are the observer and the cancel disposable.

It is intended for the constant `SignalProducer`s, as in #486, which synchronously emits all events.

It represents a new type of `SignalProducer` that **does not use `Signal`** unless it is requested via `make()`.

### Pilot public APIs

1. `map`, `mapError`, `filter` and `filterMap`.

1. Various `SignalProducer.init` overloads that accept a constant.

### Results

| Test | Sequence<br>(32 items) | Value | SequenceMapFilter |
| --- | --- | --- | --- |
| `master` | avg 30296 ns<br>min 26426 ns | avg 8910 ns<br>min 8171 ns | avg 61555 ns<br>min 50378 ns |
| This PR | avg 14073 ns<br>min 13079 ns | avg 839 ns<br>min 739 ns | avg 8304 ns<br>min 7198 ns | 
| Speedup | ~2.15x | ~10.5x | ~7.5x |
| `master`<br>WMO | avg 29333 ns<br>min 26136 ns | avg 8613 ns<br>min 7169 ns  | 31340 ns<br>min 27497 ns  |
| This PR<br>WMO | avg 13591 ns<br>min 11382 ns | avg 388 ns<br>min 325 ns | avg 5923 ns<br>min 5002 ns |
| Speedup<br >WMO | ~2.15x | ~22.00x | ~5.29x |

[Benchmark Source](https://gist.github.com/andersio/1aae47e6e9173b5dc6e8609356668ec1)

#### Checklist
- [ ] Updated CHANGELOG.md.