@andersio 2017-07-17
Heh, the compiler gets stuck.