@andersio 2017-08-15
@mdiep 

For `SignalProducer(value:)`, the allocation of a non-atomic `SimpleDisposable` contributes 29% overhead.

```
@testValue(): avg 546 ns; min 475 ns
@testValue(): avg 388 ns; min 325 ns
```