@andersio 2017-08-15
The cost of sequence is slightly up with the implementation being corrected in https://github.com/ReactiveCocoa/ReactiveSwift/pull/487/commits/56131fc38db512af012515c80c750d5308cc582f, but the difference is trivial given its scale.

All constant producers except the sequence one still uses `NopDisposable`, which is now always disposed.

```
@testEventTransformingCoreMapFilter(): avg 6091 ns; min 5250 ns
@testSequence(): avg 13277 ns; min 11720 ns
@testValue(): avg 369 ns; min 317 ns
```